Revise controller
