+++ featurebyte/routes/dimension_table/controller.py
@@ -3,13 +3,10 @@ DimensionTable API route controller
 """
 from __future__ import annotations
 
-from typing import Any, List, Tuple
-
 from bson import ObjectId
 
 from featurebyte.enum import SemanticType
 from featurebyte.models.dimension_table import DimensionTableModel
-from featurebyte.models.persistent import QueryFilter
 from featurebyte.routes.common.base_table import BaseTableDocumentController
 from featurebyte.schema.dimension_table import DimensionTableList, DimensionTableServiceUpdate
 from featurebyte.schema.info import DimensionTableInfo
@@ -41,14 +38,18 @@ class DimensionTableController(
         dimension_table_service: TableDocumentService,
         table_facade_service: TableFacadeService,
         semantic_service: SemanticService,
-        table_info_service: TableInfoService,
         entity_service: EntityService,
         feature_service: FeatureService,
+        table_info_service: TableInfoService,
     ):
-        super().__init__(dimension_table_service, table_facade_service, semantic_service)
+        super().__init__(
+            service=dimension_table_service,
+            table_facade_service=table_facade_service,
+            semantic_service=semantic_service,
+            entity_service=entity_service,
+            feature_service=feature_service,
+        )
         self.table_info_service = table_info_service
-        self.entity_service = entity_service
-        self.feature_service = feature_service
 
     async def get_info(self, document_id: ObjectId, verbose: bool) -> DimensionTableInfo:
         """
@@ -73,11 +74,3 @@ class DimensionTableController(
             **table_dict,
             dimension_id_column=dimension_table.dimension_id_column,
         )
-
-    async def service_and_query_pairs_for_checking_reference(
-        self, document_id: ObjectId
-    ) -> List[Tuple[Any, QueryFilter]]:
-        return [
-            (self.entity_service, {"primary_table_ids": document_id}),
-            (self.feature_service, {"table_ids": document_id}),
-        ]
