+++ featurebyte/routes/event_table/controller.py
@@ -45,17 +45,21 @@ class EventTableController(
         event_table_service: TableDocumentService,
         table_facade_service: TableFacadeService,
         semantic_service: SemanticService,
-        table_info_service: TableInfoService,
         entity_service: EntityService,
-        item_table_service: ItemTableService,
         feature_service: FeatureService,
+        table_info_service: TableInfoService,
+        item_table_service: ItemTableService,
         feature_job_setting_analysis_service: FeatureJobSettingAnalysisService,
     ):
-        super().__init__(event_table_service, table_facade_service, semantic_service)
+        super().__init__(
+            service=event_table_service,
+            table_facade_service=table_facade_service,
+            semantic_service=semantic_service,
+            entity_service=entity_service,
+            feature_service=feature_service,
+        )
         self.table_info_service = table_info_service
-        self.entity_service = entity_service
         self.item_table_service = item_table_service
-        self.feature_service = feature_service
         self.feature_job_setting_analysis_service = feature_job_setting_analysis_service
 
     async def get_info(self, document_id: ObjectId, verbose: bool) -> EventTableInfo:
@@ -87,9 +91,9 @@ class EventTableController(
     async def service_and_query_pairs_for_checking_reference(
         self, document_id: ObjectId
     ) -> List[Tuple[Any, QueryFilter]]:
-        return [
-            (self.entity_service, {"primary_table_ids": document_id}),
+        return await super().service_and_query_pairs_for_checking_reference(
+            document_id=document_id
+        ) + [
             (self.item_table_service, {"event_table_id": document_id}),
-            (self.feature_service, {"table_ids": document_id}),
             (self.feature_job_setting_analysis_service, {"event_table_id": document_id}),
         ]
