+++ featurebyte/routes/item_table/controller.py
@@ -3,13 +3,10 @@ ItemTable API route controller
 """
 from __future__ import annotations
 
-from typing import Any, List, Tuple
-
 from bson import ObjectId
 
 from featurebyte.enum import SemanticType
 from featurebyte.models.item_table import ItemTableModel
-from featurebyte.models.persistent import QueryFilter
 from featurebyte.routes.common.base_table import BaseTableDocumentController
 from featurebyte.schema.info import ItemTableInfo
 from featurebyte.schema.item_table import ItemTableList, ItemTableServiceUpdate
@@ -43,16 +40,20 @@ class ItemTableController(
         item_table_service: TableDocumentService,
         table_facade_service: TableFacadeService,
         semantic_service: SemanticService,
-        table_info_service: TableInfoService,
-        event_table_service: EventTableService,
         entity_service: EntityService,
         feature_service: FeatureService,
+        table_info_service: TableInfoService,
+        event_table_service: EventTableService,
     ):
-        super().__init__(item_table_service, table_facade_service, semantic_service)
+        super().__init__(
+            service=item_table_service,
+            table_facade_service=table_facade_service,
+            semantic_service=semantic_service,
+            entity_service=entity_service,
+            feature_service=feature_service,
+        )
         self.table_info_service = table_info_service
         self.event_table_service = event_table_service
-        self.entity_service = entity_service
-        self.feature_service = feature_service
 
     async def get_info(self, document_id: ObjectId, verbose: bool) -> ItemTableInfo:
         """
@@ -82,11 +83,3 @@ class ItemTableController(
             item_id_column=item_table.item_id_column,
             event_table_name=event_table.name,
         )
-
-    async def service_and_query_pairs_for_checking_reference(
-        self, document_id: ObjectId
-    ) -> List[Tuple[Any, QueryFilter]]:
-        return [
-            (self.entity_service, {"primary_table_ids": document_id}),
-            (self.feature_service, {"table_ids": document_id}),
-        ]
