+++ featurebyte/routes/scd_table/controller.py
@@ -3,12 +3,9 @@ SCDTable API route controller
 """
 from __future__ import annotations
 
-from typing import Any, List, Tuple
-
 from bson import ObjectId
 
 from featurebyte.enum import SemanticType
-from featurebyte.models.persistent import QueryFilter
 from featurebyte.models.scd_table import SCDTableModel
 from featurebyte.routes.common.base_table import BaseTableDocumentController
 from featurebyte.schema.info import SCDTableInfo
@@ -40,17 +37,21 @@ class SCDTableController(BaseTableDocumentController[SCDTableModel, SCDTableServ
 
     def __init__(
         self,
-        service: TableDocumentService,
+        scd_table_service: TableDocumentService,
         table_facade_service: TableFacadeService,
         semantic_service: SemanticService,
         table_info_service: TableInfoService,
         entity_service: EntityService,
         feature_service: FeatureService,
     ):
-        super().__init__(service, table_facade_service, semantic_service)
+        super().__init__(
+            service=scd_table_service,
+            table_facade_service=table_facade_service,
+            semantic_service=semantic_service,
+            entity_service=entity_service,
+            feature_service=feature_service,
+        )
         self.table_info_service = table_info_service
-        self.entity_service = entity_service
-        self.feature_service = feature_service
 
     async def get_info(self, document_id: ObjectId, verbose: bool) -> SCDTableInfo:
         """
@@ -79,11 +80,3 @@ class SCDTableController(BaseTableDocumentController[SCDTableModel, SCDTableServ
             end_timestamp_column=scd_table.end_timestamp_column,
             current_flag_column=scd_table.current_flag_column,
         )
-
-    async def service_and_query_pairs_for_checking_reference(
-        self, document_id: ObjectId
-    ) -> List[Tuple[Any, QueryFilter]]:
-        return [
-            (self.entity_service, {"primary_table_ids": document_id}),
-            (self.feature_service, {"table_ids": document_id}),
-        ]
